+++ kyo-net/jvm-native/src/test/scala/kyo/net/internal/posix/HandshakeEngineFreeTest.scala
@@ -60,7 +60,7 @@ class HandshakeEngineFreeTest extends Test:
       * returned number equals the count of currently-open low descriptors. Stable across leak-free iterations, climbing under an fd leak.
       */
     private def probeFd(sockets: SocketBindings, shim: PosixShimBindings)(using AllowUnsafe): Int =
-        val fd = sockets.socket(PosixConstants.AF_INET, PosixConstants.SOCK_STREAM, 0).value
+        val fd = sockets.socket(PosixConstants.AF_INET, PosixConstants.SOCK_STREAM, 0).value.toInt
         discard(shim.kyo_posix_close(fd))
         fd
     end probeFd
@@ -249,7 +249,7 @@ class HandshakeEngineFreeTest extends Test:
     /** Build a connected loopback socket pair (both non-blocking, as the transport's readiness model requires); returns (clientFd, peerFd). */
     private def loopbackPair()(using Frame, kyo.test.AssertScope): (Int, Int) < Async =
         val sock   = Ffi.load[SocketBindings]
-        val server = sock.socket(PosixConstants.AF_INET, PosixConstants.SOCK_STREAM, 0).value
+        val server = sock.socket(PosixConstants.AF_INET, PosixConstants.SOCK_STREAM, 0).value.toInt
         val (a, l) = SockAddr.encodeInet4(PosixConstants.AF_INET, "127.0.0.1", 0).getOrElse(fail("encode failed"))
         Sync.ensure(Sync.defer(a.close())) {
             assert(sock.bind(server, a, l).value == 0)
@@ -264,7 +264,7 @@ class HandshakeEngineFreeTest extends Test:
                 finally
                     out.close()
                     ol.close()
-            val client   = sock.socket(PosixConstants.AF_INET, PosixConstants.SOCK_STREAM, 0).value
+            val client   = sock.socket(PosixConstants.AF_INET, PosixConstants.SOCK_STREAM, 0).value.toInt
             val (ca, cl) = SockAddr.encodeInet4(PosixConstants.AF_INET, "127.0.0.1", port).getOrElse(fail("encode failed"))
             val connected =
                 Sync.ensure(Sync.defer(ca.close()))(sock.connect(client, ca, cl).safe.get.map(r => assert(r.value == 0)))
@@ -273,7 +273,7 @@ class HandshakeEngineFreeTest extends Test:
                 val noLen  = Buffer.alloc[Int](1)
                 noLen.set(0, SockAddr.inet4Size)
                 Sync.ensure(Sync.defer { noAddr.close(); noLen.close() }) {
-                    sock.accept(server, noAddr, noLen).safe.get.map(_.value)
+                    sock.accept(server, noAddr, noLen).safe.get.map(_.value.toInt)
                 }.map { accepted =>
                     val shim = Ffi.load[PosixShimBindings]
                     assert(shim.kyo_posix_set_nonblocking(client) == 0, "set client non-blocking")
