+++ kyo-net/jvm-native/src/test/scala/kyo/net/internal/posix/IoUringDriverTest.scala
@@ -828,7 +828,7 @@ class IoUringDriverTest extends Test:
 
     /** Bind+listen on an ephemeral 127.0.0.1 port and return the listening server fd. */
     private def listenOnly()(using Frame, kyo.test.AssertScope): Int < Async =
-        val server = sock.socket(PosixConstants.AF_INET, PosixConstants.SOCK_STREAM, 0).value
+        val server = sock.socket(PosixConstants.AF_INET, PosixConstants.SOCK_STREAM, 0).value.toInt
         val (a, l) = SockAddr.encodeInet4(PosixConstants.AF_INET, "127.0.0.1", 0).getOrElse(fail("encode failed"))
         Sync.ensure(Sync.defer(a.close())) {
             assert(sock.bind(server, a, l).value == 0)
@@ -849,7 +849,7 @@ class IoUringDriverTest extends Test:
             finally
                 out.close()
                 ol.close()
-        val client   = sock.socket(PosixConstants.AF_INET, PosixConstants.SOCK_STREAM, 0).value
+        val client   = sock.socket(PosixConstants.AF_INET, PosixConstants.SOCK_STREAM, 0).value.toInt
         val (ca, cl) = SockAddr.encodeInet4(PosixConstants.AF_INET, "127.0.0.1", port).getOrElse(fail("encode failed"))
         Sync.ensure(Sync.defer(ca.close()))(sock.connect(client, ca, cl).safe.get.map(r => assert(r.value == 0))).map(_ => client)
     end connectTo
